+++ lib/reduced_bloc.dart
@@ -2,5 +2,5 @@
 
 library reduced_bloc;
 
-export 'src/bloc_reducible.dart';
-export 'src/bloc_wrapper.dart';
+export 'src/bloc_store.dart';
+export 'src/bloc_widgets.dart';