+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.1.0
 
 * initial release.
+
+## 0.2.0
+
+* upgrade from reduced 0.1.0 to reduced 0.2.0
+* replace function wrapWithProvider with widget class ReducedProvider
+* replace function wrapWithConsumer with widget class ReducedConsumer
+* replace function wrapWithScope with widget class ReducedScope